% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_electrode.R
\name{calc_cytf}
\alias{calc_cytf}
\title{Calculate cyt f content}
\usage{
calc_cytf(pmax, psii, coefs = c(0.0217, 0.004))
}
\arguments{
\item{pmax}{maximum gross photosynthetic O2 evolution rate, unit: micro mol m^-2 s^-1}

\item{psii}{functional PSII content, unit: micro mol m^-2.}

\item{coefs}{coefficients for the calculation reported by Dwyer et al. (2012).}
}
\value{
a numeric vector
}
\description{
.
This function calculates cytochrome f content according to Dwyer et al. (2012; doi:10.1093/jxb/ers156).
}
