% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_electrode.R
\name{calc_oxy}
\alias{calc_oxy}
\title{Calculate Pmax, functional PSII content, and cyt f content}
\usage{
calc_oxy(calb, positive_pmax, negative_pmax, positive_psii, negative_psii,
  amp_pmax = 5, amp_psii = 20, leaf_area, hz = 10, e_per_o2 = 4)
}
\arguments{
\item{calb}{signal amplitude corresponding to 1 mL O2 which was injected for the calibration, unit: grid.}

\item{positive_pmax}{a slope corresponding to net O2 evolution rate under repeatitive single-turnover flashes and weak far-red light, unit: grid min^-1.}

\item{negative_pmax}{a slope corresponding to O2 uptake rate under weak far-red light, unit: grid min^-1.}

\item{positive_psii}{a slope corresponding to net O2 evolution rate under repeatitive single-turnover flashes and weak far-red light, unit: grid min^-1.}

\item{negative_psii}{a slope corresponding to O2 uptake rate under weak far-red light, unit: grid min^-1.}

\item{amp_pmax}{an expansion coefficient, no-unit.}

\item{amp_psii}{an expansion coefficient, no-unit.}

\item{leaf_area}{an area of the enclosed leaf in the chamber, unit: m^2.}

\item{hz}{a frequency of single-turnover flash, unit: s-1, default: 10.}

\item{e_per_o2}{a number of electron consumed in PSII to evolve one O2, unit: mol mol^-1, default: 4.}

\item{artefact}{a factor to correct non-photosynthetic signal, i.e. heating by irradiance, unit: grid min^-1.}
}
\value{
a lst
}
\description{
.
This function calculates three key parameters.
}
