% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_electrode.R
\name{calc_pgmax_o2}
\alias{calc_pgmax_o2}
\title{Calculate Pmax}
\usage{
calc_pgmax_o2(calb, posisitve, negative, amp, leaf_area)
}
\arguments{
\item{calb}{signal amplitude corresponding to 1 mL O2 which was injected for the calibration, unit: grid.}

\item{negative}{a slope corresponding to dark O2 uptake rate, unit: grid min^-1.}

\item{amp}{an expansion coefficient, no-unit.}

\item{leaf_area}{an area of the enclosed leaf in the chamber, unit: m^2.}

\item{positive}{a slope corresponding to net O2 evolution rate at saturating irradiance, unit: grid min^-1.}
}
\value{
a numeric vector
}
\description{
.
This function calculates maximum gross photosynthetic rates (Pmax) from net O2 evolution and dark O2 uptake rates.
}
