% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_lambda.R
\name{read_field_lambda}
\alias{read_field_lambda}
\title{Read field-lambda data}
\usage{
read_field_lambda(file_path, unit_select = c("photon", "irradiance")[1],
  col_types = cols(), ..., .verbose = F)
}
\value{
a tibble
}
\description{
Read field-lambda data
}
